fix: refactoring api client-side file structure #1
