+++ Client-side/StreamingApp/src/API/User.routes.ts
