+++ Client-side/StreamingApp/src/vite-env.d.ts
@@ -1,0 +1,9 @@
+/// <reference types="vite/client" />
+interface ImportMetaEnv {
+    readonly VITE_BASE_URL: string
+    // more env variables...
+}
+  
+interface ImportMeta {
+readonly env: ImportMetaEnv
+}
